Organize headers and sources
