+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "flashmatch.hpp"
+#include "flashmatch/flashmatch.hpp"
 
 int main(int argc, char *argv[]) {
   return flashmatch_main(argc, argv);
