+++ tests/test_lock_free_queue.cpp
@@ -1,4 +1,4 @@
-#include "lock_free_queue.hpp"
+#include "flashmatch/lock_free_queue.hpp"
 #include <gtest/gtest.h>
 #include <stdexcept>
 #include <thread>
