+++ tests/test_main.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include "flashmatch.hpp"
+#include "flashmatch/flashmatch.hpp"
 
 TEST(MainTest, MainReturnsZero) {
   EXPECT_EQ(0, flashmatch_main(0, nullptr));
